- Master merged in
- Lot of fixes to allow the application to run
- SIGSEGVs remaining and buttons too smalls

Conflicts:
	AABB_tree/demo/AABB_tree/Scene.cpp
	AABB_tree/demo/AABB_tree/Scene.h
	AABB_tree/demo/AABB_tree/Viewer.cpp
	GraphicsView/src/CGAL_Qt5/CMakeLists.txt
	Installation/CMakeLists.txt
	Polyhedron/demo/Polyhedron/CGAL_demo/Viewer_interface.h
	Polyhedron/demo/Polyhedron/CMakeLists.txt
	Polyhedron/demo/Polyhedron/GlSplat/GlSplat.h
	Polyhedron/demo/Polyhedron/GlSplat/Shader.h
	Polyhedron/demo/Polyhedron/MainWindow.cpp
	Polyhedron/demo/Polyhedron/MainWindow.ui
	Polyhedron/demo/Polyhedron/Polyhedron_demo_cut_plugin.cpp
	Polyhedron/demo/Polyhedron/Polyhedron_demo_point_inside_polyhedron_plugin.cpp
	Polyhedron/demo/Polyhedron/Scene.cpp
	Polyhedron/demo/Polyhedron/Scene.h
	Polyhedron/demo/Polyhedron/Scene_c2t3_item.h
	Polyhedron/demo/Polyhedron/Scene_combinatorial_map_item.cpp
	Polyhedron/demo/Polyhedron/Scene_draw_interface.h
	Polyhedron/demo/Polyhedron/Scene_edit_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Scene_edit_polyhedron_item.h
	Polyhedron/demo/Polyhedron/Scene_item.cpp
	Polyhedron/demo/Polyhedron/Scene_item.h
	Polyhedron/demo/Polyhedron/Scene_nef_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Scene_nef_polyhedron_item.h
	Polyhedron/demo/Polyhedron/Scene_nef_rendering.cpp
	Polyhedron/demo/Polyhedron/Scene_nef_rendering.h
	Polyhedron/demo/Polyhedron/Scene_points_with_normal_item.cpp
	Polyhedron/demo/Polyhedron/Scene_points_with_normal_item.h
	Polyhedron/demo/Polyhedron/Scene_polygon_soup_item.h
	Polyhedron/demo/Polyhedron/Scene_polyhedron_item.cpp
	Polyhedron/demo/Polyhedron/Scene_polylines_item.cpp
	Polyhedron/demo/Polyhedron/Viewer.cpp
	Polyhedron/demo/Polyhedron/include/Point_set_3.h
	Polyhedron/demo/Polyhedron/resources/shader_with_textured_edges.f
	Polyhedron/demo/Polyhedron/resources/shader_without_light.v

+++ Polyhedron/demo/Polyhedron/Scene_combinatorial_map_item.h
@@ -73,10 +73,10 @@ private:
     template <class Predicate> void export_as_polyhedron(Predicate,const QString&) const;
 
 
-   mutable std::vector<double> positions_lines;
-   mutable std::vector<double> positions_points;
-   mutable std::vector<double> positions_facets;
-   mutable std::vector<double> normals;
+   mutable std::vector<float> positions_lines;
+   mutable std::vector<float> positions_points;
+   mutable std::vector<float> positions_facets;
+   mutable std::vector<float> normals;
    mutable std::size_t nb_lines;
    mutable std::size_t nb_points;
    mutable std::size_t nb_facets;
