+++ Polyhedron/demo/Polyhedron/Scene_nef_polyhedron_item.h
@@ -75,13 +75,13 @@ private:
   Nef_polyhedron* nef_poly;
 
 
-  mutable std::vector<double> positions_lines;
-  mutable std::vector<double> positions_facets;
-  mutable std::vector<double> positions_points;
-  mutable std::vector<double> normals;
-  mutable std::vector<double> color_lines;
-  mutable std::vector<double> color_facets;
-  mutable std::vector<double> color_points;
+  mutable std::vector<float> positions_lines;
+  mutable std::vector<float> positions_facets;
+  mutable std::vector<float> positions_points;
+  mutable std::vector<float> normals;
+  mutable std::vector<float> color_lines;
+  mutable std::vector<float> color_facets;
+  mutable std::vector<float> color_points;
   mutable std::size_t nb_points;
   mutable std::size_t nb_lines;
   mutable std::size_t nb_facets;
